AE - fix publish after background load
