+++ pype/lib/__init__.py
@@ -25,7 +25,12 @@ from .applications import (
     _subprocess
 )
 
-from .plugin_tools import filter_pyblish_plugins, source_hash
+from .plugin_tools import (
+    filter_pyblish_plugins,
+    source_hash,
+    get_unique_layer_name,
+    get_background_layers
+)
 
 from .path_tools import (
     version_up,
@@ -57,6 +62,8 @@ __all__ = [
     "ApplicationAction",
 
     "filter_pyblish_plugins",
+    "get_unique_layer_name",
+    "get_background_layers",
 
     "version_up",
     "get_version_from_path",
