+++ pype/plugins/aftereffects/load/load_background.py
@@ -2,7 +2,7 @@ import re
 
 from avalon import api, aftereffects
 
-from pype.plugins.lib import get_background_layers, get_unique_layer_name
+from pype.lib import get_background_layers, get_unique_layer_name
 
 stub = aftereffects.stub()
 
@@ -51,6 +51,7 @@ class BackgroundLoader(api.Loader):
     def update(self, container, representation):
         """ Switch asset or change version """
         context = representation.get("context", {})
+        layer = container.pop("layer")
 
         # without iterator number (_001, 002...)
         namespace_from_container = re.sub(r'_\d{3}$', '',
