+++ pype/plugins/aftereffects/publish/collect_render.py
@@ -33,12 +33,16 @@ class CollectAERender(abstract_collect_render.AbstractCollectRender):
 
         compositions = aftereffects.stub().get_items(True)
         compositions_by_id = {item.id: item for item in compositions}
-        for item_id, inst in aftereffects.stub().get_metadata().items():
+        for inst in aftereffects.stub().get_metadata():
             schema = inst.get('schema')
             # loaded asset container skip it
             if schema and 'container' in schema:
                 continue
 
+            if not inst["members"]:
+                raise ValueError("Couldn't find id, unable to publish. " +
+                                 "Please recreate instance.")
+            item_id = inst["members"][0]
             work_area_info = aftereffects.stub().get_work_area(int(item_id))
             frameStart = work_area_info.workAreaStart
